+++ asusd/src/ctrl_anime/mod.rs
@@ -73,6 +73,20 @@ impl CtrlAnime {
             return Err(RogError::Anime(AnimeError::NoDevice));
         };
 
+        // TODO: something better to set wakeups disabled
+        if matches!(node, Node::Usb(_)) {
+            if let Ok(mut enumerator) = udev::Enumerator::new() {
+                enumerator.match_subsystem("usb").ok();
+                enumerator.match_attribute("idProduct", "193b").ok();
+
+                if let Ok(mut enumer) = enumerator.scan_devices() {
+                    if let Some(mut dev) = enumer.next() {
+                        dev.set_attribute_value("power/wakeup", "disabled").ok();
+                    }
+                }
+            }
+        }
+
         let mut anime_type = get_anime_type()?;
         if let AnimeType::Unknown = anime_type {
             if let Some(model) = config.model_override {